+++ src/views/func/workflow/index.vue
@@ -1,0 +1,27 @@
+<template>
+   <div class="page-container">
+        workflow
+   </div>
+</template>
+
+<script setup name="funcWorkflow">
+const { proxy } = getCurrentInstance() // vue 实例
+const router = useRouter() // router 实例
+const route = useRoute() // 路由参数
+const store = useStore() // vuex 实例
+
+
+// 定义响应式数据>
+const state = reactive({
+
+})
+
+
+// 组件挂载后，此方法执行后，页面显示
+onMounted(() => {
+
+})
+</script>
+
+<style lang='scss' scoped>
+</style>